+++ CHANGES.md
@@ -18,8 +18,8 @@
 - [UPDATE] libwebrtc を m137.7151.0.0 にあげる
   - `rtc::revive::` を　`webrtc::revive::` に変更する
   - BasicPortAllocator の引数追加に伴い、`config_.env` を追加する
-  - m137 で `rtc_base/third_party/base64/base64.h` が削除されたため、websocket.cpp で `absl/strings/escaping.h` をインクルードするように変更する
-  - `webrtc::Base64::Encode` から `absl::Base64Escape` に変更する
+  - m137 で `rtc_base/third_party/base64/base64.h` が削除されたため、websocket.cpp で `rtc_base/base64.h` をインクルードするように変更する
+  - `webrtc::Base64::Encode` から `webrtc::Base64Encode` に変更する
   - @miosakuma @torikizi
 
 ### misc